Add progress bars to workout card and update the form
